Unify nns ut
